requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
chardet==4.0.0
decorator==4.4.2
discord.py==1.7.1
idna==3.1
Pillow==8.2.0
pycodestyle==2.7.0
pydantic==1.8.1
pyparsing==2.4.5
python-dateutil==2.8.1
python-dotenv==0.17.0

